# src/docs/swagger.yaml
basePath: /v1
definitions:
...
  main.ConfirmOtpRequest:
    properties:
      code:
        type: string
      user_id:
        type: string
    required:
    - code
    - user_id
    type: object
...
  main.RequestRecoveryPassword:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  main.RequestRecoveryPasswordResponse:
    properties:
      success:
        type: boolean
    type: object
  main.RequestSignUpRequest:
    properties:
      dailing_code:
        type: string
      phone_number:
        type: string
    required:
    - dailing_code
    - phone_number
    type: object
  main.RequestSignUpResponse:
    properties:
      expires_at:
        type: string
      max_attempts:
        type: integer
      session:
        type: string
      user_id:
        type: string
    type: object
  main.ResetPasswordRequest:
    properties:
      code:
        type: string
      email:
        type: string
      password:
        type: string
    required:
    - code
    - email
    - password
    type: object
  main.ResetPasswordResponse:
    properties:
      success:
        type: boolean
    type: object
  main.SignUpByEmailRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  main.SignUpByEmailResponse:
    properties:
      user_id:
        type: string
    type: object
...
  main.VerifyOTPByPhoneRequest:
    properties:
      code:
        type: string
      session:
        type: string
      user_id:
        type: string
    required:
    - code
    - session
    - user_id
    type: object
  main.VerifyOTPByPhoneResponse:
    properties:
      expires_at:
        type: string
      max_attempts:
        type: integer
      session:
        type: string
      user_id:
        type: string
    type: object
  utils.Response-any:
    properties:
      data: {}
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  utils.Response-main_RequestRecoveryPasswordResponse:
    properties:
      data:
        $ref: '#/definitions/main.RequestRecoveryPasswordResponse'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
  utils.Response-main_RequestSignUpResponse:
    properties:
      data:
        $ref: '#/definitions/main.RequestSignUpResponse'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
  utils.Response-main_ResetPasswordResponse:
    properties:
      data:
        $ref: '#/definitions/main.ResetPasswordResponse'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
  utils.Response-main_SignUpByEmailResponse:
    properties:
      data:
        $ref: '#/definitions/main.SignUpByEmailResponse'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  utils.Response-main_VerifyOTPByPhoneResponse:
    properties:
      data:
        $ref: '#/definitions/main.VerifyOTPByPhoneResponse'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
host: api.smartgo.com
info:
  contact: {}
  description: SmartGo API
  title: SmartGo API
  version: "1.0"
paths:
  /auth:
    patch:
      consumes:
      - application/json
      description: |-
        This endpoint verifies a OTP by email.
        Verify the code given a userID and a code. The code is the code that the user received in the email.
        Also, the email will be verified as true in provider (cognito) and database.
      parameters:
      - description: Request body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.ConfirmOtpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Login with email and password
      tags:
      - security
    post:
      consumes:
      - application/json
      description: |-
        This endpoint signs up a user by email and password.
        The user will be created in the database and inderity provider (cognito) if it doesn't exist.
        If the user exists, an error will be returned.
        A mail will be sent to the user to verify the email.
        If the user exists but the email is not verified, a new OTP will be sent to the user.
      parameters:
      - description: Request body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.SignUpByEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-main_SignUpByEmailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Sign up by email and password
      tags:
      - security
  /auth/otp:
    patch:
      consumes:
      - application/json
      description: |-
        This endpoint verifies a OTP by phone number.
        If the OTP is valid, the user will be logged in and return the token entity.
        If the OTP is invalid, the user will be logged in and return the login challenge entity.
        The login challenge entity contains the session and the userID with which the user can try to verify the OTP again.
      parameters:
      - description: Request body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.VerifyOTPByPhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-main_VerifyOTPByPhoneResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Verify OTP from phone
      tags:
      - security
    post:
      consumes:
      - application/json
      description: |-
        Try to sign up or login by phone number.
        The user will be created in the database and inderity provider (cognito) if it doesn't exist.
        If the user exists, the code will be sent to the phone number.
        The user will be created without password and return the session.
        The session is specific context for the user to login in the cognito user pool. It is used to verify the user's phone number.
        For others implementations, this session could be not needed.
        The user ID (uuid type) is the user's ID in the database and also the username in the cognito user pool.
      parameters:
      - description: Request body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.RequestSignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-main_RequestSignUpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Request sign up / login by phone
      tags:
      - security
  /auth/recovery-password:
    patch:
      consumes:
      - application/json
      description: Reset password. If the code, password and email are valid, the
        password will be reset.
      parameters:
      - description: Reset password request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-main_ResetPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Reset password
      tags:
      - security
    post:
      consumes:
      - application/json
      description: |-
        Request password recovery. If the email is found, the user will receive an email with a code to reset the password.
        In otherwise, the user will receive an error.
      parameters:
      - description: Request password recovery
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.RequestRecoveryPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-main_RequestRecoveryPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Request password recovery
      tags:
      - security
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"